extract local file discovery into method
it seems like this gigantic loop resets and repopulates the list of
local files in a few different circumstances
I think that this is working nicely for mouse clicks, however for
keyboard presses this doesn't seem to happen.
I disovered a commented-out else statement near the bottom of the loop
which seemed to re-initialise the file list at some point for this
keypress case.
I've tried to reinstate the logic in a working fashion